Resolução de problemas no repositorio

+++ app/src/main/java/com/tinderdog/repository/helper/sqlite/LocalDBHelper.java
@@ -23,23 +23,23 @@ public class LocalDBHelper extends SQLiteOpenHelper {
 
 
     private LocalDBHelper() {
-        super(TinderDogApp.getContext(), "tinderdog.db", null, 1);
+        super(TinderDogApp.getContext(), "tindedog", null, 2);
     }
 
     @Override
     public void onCreate(SQLiteDatabase db) {
-        LoggerWrapper.log("DEBE CRIOOOOOOOOOOOOOOOOOOOOOO");
-        String sql =
-                "CREATE TABLE dogs (id INTEGER PRIMARY KEY AUTOINCREMENT, owner_id INTEGER REFERENCES users (id) ON DELETE CASCADE ON UPDATE NO ACTION NOT NULL, name VARCHAR (180), age DOUBLE, gait VARCHAR (180), color VARCHAR (180));"
-                +"CREATE TABLE user_adresses (user_id INTEGER REFERENCES users (id) ON DELETE CASCADE ON UPDATE NO ACTION UNIQUE NOT NULL PRIMARY KEY, cep VARCHAR (180), logradouro VARCHAR (180), bairro VARCHAR (180), cidade VARCHAR (180), estado VARCHAR (180));"
-                +"CREATE TABLE users (id INTEGER PRIMARY KEY AUTOINCREMENT, name VARCHAR (180) NOT NULL, email VARCHAR (180), password VARCHAR (128), cpf VARCHAR (11) UNIQUE, birth_date DATE);";
-        db.execSQL(sql);
+        db.execSQL("CREATE TABLE users (id INTEGER PRIMARY KEY AUTOINCREMENT, name VARCHAR (180) NOT NULL, email VARCHAR (180), password VARCHAR (128), cpf VARCHAR (11) UNIQUE, birth_date DATE);");
+        db.execSQL("CREATE TABLE user_adresses (user_id INTEGER REFERENCES users (id) ON DELETE CASCADE ON UPDATE NO ACTION UNIQUE NOT NULL PRIMARY KEY, cep VARCHAR (180), logradouro VARCHAR (180), bairro VARCHAR (180), cidade VARCHAR (180), estado VARCHAR (180));");
+        db.execSQL("CREATE TABLE dogs (id INTEGER PRIMARY KEY AUTOINCREMENT, owner_id INTEGER REFERENCES users (id) ON DELETE CASCADE ON UPDATE NO ACTION NOT NULL, name VARCHAR (180), age DOUBLE, gait VARCHAR (180), color VARCHAR (180));");
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int i, int i1) {
         Cursor c = db.rawQuery("SELECT name FROM sqlite_master WHERE type='table'", null);
         while (c.moveToNext()) {
+            if (c.getString(0).equals("sqlite_sequence")){
+                continue;
+            }
             db.execSQL("DROP TABLE IF EXISTS " + c.getString(0));
         }
         c.close();
